+++ .gitignore
@@ -4,7 +4,6 @@
 .ipynb_checkpoints/
 .mypy_cache/
 .pytest_cache/
-.tox/
 .venv/
 .vscode/
 *.egg-info/
