+++ setup.cfg
@@ -35,6 +35,8 @@ flake8-ignore =
     */types.py:F401
     */models_autogenerated.py E501
     *models_auto.py E501
+    * E721
+    * W503
 
 [rstcheck]
 ignore_messages=(No role entry for "samp")|(Duplicate implicit target name: )|(Hyperlink target .* is not referenced.)
