--- tox.ini
@@ -1,8 +1,0 @@
-# content of: tox.ini , put in same dir as setup.py
-[tox]
-envlist = py36,py37,py38
-
-[testenv]
-extras = test
-commands =
-    pytest
